Cleanup

+++ test/get_remote_test.cc
@@ -1,5 +1,6 @@
 #include "doctest.h"
 
+#include <boost/filesystem/directory.hpp>
 #include <system_error>
 
 #include <boost/filesystem.hpp>
@@ -14,10 +15,13 @@ namespace fs = boost::filesystem;
 using pkg::exec;
 using pkg::get_remote;
 
-auto const test_repo = fs::path{TEST_EXECUTION_DIR} / "test" / "git_test_repo";
+TEST_CASE("get_remote") {
+  auto const test_repo =
+      fs::path{TEST_EXECUTION_DIR} / "test" / "git_test_repo";
 
-void setup() {
-  fs::remove_all(test_repo);
+  if (fs::exists(test_repo)) {
+    fs::remove_all(test_repo);
+  }
   fs::create_directory(test_repo);
   exec(test_repo, "git init");
   exec(test_repo, "git remote add origin git@github.com:foo/foo.git");
@@ -27,10 +31,6 @@ void setup() {
   exec(test_repo, "git remote add remoteD ssh://git@github.com:baz/baz.git");
   exec(test_repo, "git remote add remoteE git@github.com:zab/zab.git");
   exec(test_repo, "git remote set-url --push remoteE read-only");
-}
-
-TEST_CASE("get_remote") {
-  setup();
 
   CHECK("remoteA" == get_remote(test_repo, "git@github.com:bar/bar.git"));
   CHECK("remoteD" == get_remote(test_repo, "ssh://git@github.com:baz/baz.git"));
